### 2018-06-10 #769 @whedon
[PRE REVIEW]: Python Implementation of Codon Adaptation Index
**Submitting author:** @benjamin-lee (<a href="http://orcid.org/0000-0002-7133-8397">Benjamin Lee</a>)
**Repository:** <a href="https://github.com/Benjamin-Lee/CodonAdaptationIndex" target ="_blank">https://github.com/Benjamin-Lee/CodonAdaptationIndex</a>
**Version:** v1.0.0
