### 2018-06-11 #769 @Benjamin-Lee
Just a question/comment: "license" is spelled as "licence" in the section heading and "license" in the section body. Is this deliberate?
