### 2018-06-11 #769 @arfon
> Just a question/comment: "license" is spelled as "licence" in the section heading and "license" in the section body. Is this deliberate?

Nope, that's just a typo. Thanks for catching that!
