### 2018-06-11 #769 @Benjamin-Lee
No problem. Happy to help!

Another question: if I make minor changes to my paper (changing word order), how do I update the proof to reflect those changes?
