### 2018-06-12 aa765eb
Reword paper intro paragraph
