### 2018-06-25 #769 @arfon
@mgymrek @george-githinji - would one of you be willing to edit this submission for JOSS?
