### 2018-06-27 6e7576d
Paper edits
