### 2018-06-27 5501180
Moved citation inside sentence
