### 2018-07-27 171d62d
PEP fix for setup.py
