### 2018-08-19 #769 @arfon
@Benjamin-Lee - please take a look at this list of potential reviewers and suggest some people who would be good potential reviewers: https://bit.ly/joss-reviewers
