### 2018-08-22 8dda044
Updates to the paper and v1.0.1
