### 2018-08-22 928dd02
Added my middle name to the paper
