### 2018-08-27 7f76002
Added codecov to travis
