### 2018-08-27 6b6ef2e
Added codecov to travis
