+++ .travis.yml
@@ -5,6 +5,9 @@ python:
   - "3.6"
 install:
     - pip install -e .
+    - pip install codecov pytest-cov
 script:
-    - pytest
+    - pytest --cov=freqgen
 cache: pip
+after_success:
+    - codecov
