+++ .travis.yml
@@ -7,7 +7,7 @@ install:
     - pip install -e .
     - pip install codecov pytest-cov
 script:
-    - pytest --cov=freqgen
+    - pytest --cov=CAI
 cache: pip
 after_success:
     - codecov
