+++ .travis.yml
@@ -7,7 +7,7 @@ install:
     - pip install -e .
     - pip install codecov pytest-cov
 script:
-    - pytest --cov=CAI
+    - py.test --cov=.
 cache: pip
 after_success:
     - codecov
