### 2018-09-16 Benjamin-Lee/CodonAdaptationIndex#2 @ctb
Some test input files would be useful for newcomers.
It would be great to provide some full examples that let a new user immediately run CAI on some real data. For my JOSS review I'm going to have to hunt down the right files before I review the functionality!
