### 2018-09-16 Benjamin-Lee/CodonAdaptationIndex#3 @ctb
More guidance on contributing / filing issues.
In the README, it would be good to provide more guidance, e.g. "open a pull request to contribute" (With a pointer to github documentation) and "use the issue tracker (link) to file an issue."
