### 2018-09-16 #905 @ctb
Finished the easy bit of the review and filed some suggestions in issues --

- [ ] track down some example input files https://github.com/Benjamin-Lee/CodonAdaptationIndex/issues/2
- [ ] add a bit more documentation about contributing and filing issues https://github.com/Benjamin-Lee/CodonAdaptationIndex/issues/3
