### 2018-09-17 9877021
Clean up the paper
