### 2018-09-17 4bbcf34
Clarify performance clain
