### 2018-09-30 #905 @ctb
Hello! Thank you for this! I think this is perfectly acceptable for publication.

The only _suggestion_ I have is that a complete real-world set of sequences be made available (along with test instructions) either as part of the repo or for download.  Specifically, under https://cai.readthedocs.io/en/latest/usage.html#basic-usage, it would be great to have a command to run that actually worked with files that were readily available. This helps users identify format issues and otherwise engage quickly with the software. But it is not essential for acceptance.

Nice work!
