### 2018-09-30 8366bc4
Fix annoying RTD issues
