### 2018-09-30 8dd8c04
Include path to example seqs
