### 2018-09-30 7803090
clarify that CAI and cai both work from the CLI
