### 2018-09-30 b9447e6
Fix incorrect italics for E coli
