### 2018-09-30 7bf8d7d
Add where to find more example sequences
