### 2018-09-30 #905 @Benjamin-Lee
I just updated the https://cai.readthedocs.io/en/latest/usage.html#basic-usage section to point out specifically that the referenced files are a part of the repository. In addition, the [example_seqs](https://github.com/Benjamin-Lee/CodonAdaptationIndex/blob/master/example_seqs/) directory on the GitHub contains six reference sets of highly expressed genes and six example CDSs.
