### 2018-09-30 #905 @Benjamin-Lee
Great! Anything else I need to get the check box for examples?

> On Sep 30, 2018, at 7:14 PM, C. Titus Brown <notifications@github.com> wrote:
>
> Excellent! Sorry I missed it in my review - I did some find commands but clearly not good ones :)
>
> --
> Titus Brown, ctbrown@ucdavis.edu<mailto:ctbrown@ucdavis.edu>
>
> On Sep 30, 2018, at 4:12 PM, Benjamin Lee <notifications@github.com<mailto:notifications@github.com>> wrote:
>
>
> I just updated the https://cai.readthedocs.io/en/latest/usage.html#basic-usage section to point out specifically that the referenced files are a part of the repository. In addition, the example_seqs<https://github.com/Benjamin-Lee/CodonAdaptationIndex/blob/master/example_seqs/> directory on the GitHub contains six reference sets of highly expressed genes and six example CDSs.
>
