### 2018-10-01 94a450a
Require the arguments to the CLI
