### 2018-10-02 #905 @Benjamin-Lee
`10.5281/zenodo.1439494` is the archived version of v1.0.2
