### 2018-10-02 #905 @Benjamin-Lee
@arfon Quick question: is there supposed to be a space between the ")" and "P" in "E. Mocarski, Ed.)PLOS Pathogens" in the citations?
