### 2018-10-03 #905 @arfon
> @arfon Quick question: is there supposed to be a space between the ")" and "P" in "E. Mocarski, Ed.)PLOS Pathogens" in the citations?

That looks like a bug in the citation style formatting. Perhaps just remove this line from your bibtex?:

https://github.com/Benjamin-Lee/CodonAdaptationIndex/blob/master/paper/paper.bib#L52
