### 2018-10-03 #905 @Benjamin-Lee
@arfon That did the trick! I looked into [whedon/resources/apa.csl](https://github.com/openjournals/whedon/blob/master/resources/apa.csl) but didn't make much progress on identifying the problem to prevent it from happening to other authors. Am I looking in the right place?

That being said, the proof looks like its good to go. Thanks!
