### 2018-10-04 #905 @arfon
> @arfon That did the trick! I looked into whedon/resources/apa.csl but didn't make much progress on identifying the problem to prevent it from happening to other authors. Am I looking in the right place?

Yeah, that's probably where. Thanks for looking!
